# ARROW-7754: [C++] Make Result<> faster