--- a/cpp/src/arrow/result.cc
+++ b/cpp/src/arrow/result.cc
@@ -27,6 +27,10 @@ namespace internal {
 
 void DieWithMessage(const std::string& msg) { ARROW_LOG(FATAL) << msg; }
 
+void InvalidValueOrDie(const Status& st) {
+  DieWithMessage(std::string("ValueOrDie called on an error: ") + st.ToString());
+}
+
 }  // namespace internal
 
 }  // namespace arrow