The main culprit for slowness was the underlying variant<> implementation.

The new implementation has comparable speed to returning Status + passing
a result-out pointer parameter.