@pitrou 2020-02-05
Benchmarks with gcc 7.4.
Before:
```
ErrorSchemeNoError                 6639 ns         6639 ns       317032 items_per_second=903.797M/s
ErrorSchemeBool                    6671 ns         6670 ns       315067 items_per_second=899.551M/s
ErrorSchemeStatus                  6691 ns         6690 ns       315033 items_per_second=896.894M/s
ErrorSchemeResult                 17563 ns        17561 ns       119493 items_per_second=341.658M/s
ErrorSchemeException               6661 ns         6660 ns       314578 items_per_second=900.89M/s
ErrorSchemeNoErrorNoInline         8055 ns         8054 ns       261182 items_per_second=745.009M/s
ErrorSchemeBoolNoInline            6966 ns         6965 ns       301415 items_per_second=861.476M/s
ErrorSchemeStatusNoInline         16561 ns        16558 ns       124946 items_per_second=362.354M/s
ErrorSchemeResultNoInline         26487 ns        26484 ns        79153 items_per_second=226.553M/s
ErrorSchemeExceptionNoInline       6713 ns         6713 ns       312970 items_per_second=893.829M/s
```
After:
```
ErrorSchemeNoError                 6625 ns         6624 ns       318577 items_per_second=905.814M/s
ErrorSchemeBool                    6627 ns         6626 ns       317117 items_per_second=905.542M/s
ErrorSchemeStatus                  6618 ns         6617 ns       317080 items_per_second=906.801M/s
ErrorSchemeResult                  6679 ns         6678 ns       315035 items_per_second=898.482M/s
ErrorSchemeException               6619 ns         6618 ns       317465 items_per_second=906.599M/s
ErrorSchemeNoErrorNoInline         7990 ns         7988 ns       261371 items_per_second=751.097M/s
ErrorSchemeBoolNoInline            6934 ns         6933 ns       303121 items_per_second=865.388M/s
ErrorSchemeStatusNoInline         16428 ns        16426 ns       127358 items_per_second=365.283M/s
ErrorSchemeResultNoInline         16525 ns        16522 ns       126809 items_per_second=363.152M/s
ErrorSchemeExceptionNoInline       7980 ns         7979 ns       263240 items_per_second=751.949M/s
```