@pitrou 2020-02-05
Benchmarks with clang 9.0.
Before:
```
ErrorSchemeNoError                 7201 ns         7200 ns       292201 items_per_second=833.287M/s
ErrorSchemeBool                    7270 ns         7269 ns       289349 items_per_second=825.452M/s
ErrorSchemeStatus                  6888 ns         6887 ns       303353 items_per_second=871.182M/s
ErrorSchemeResult                 48606 ns        48600 ns        43180 items_per_second=123.456M/s
ErrorSchemeException               7675 ns         7674 ns       273419 items_per_second=781.85M/s
ErrorSchemeNoErrorNoInline         8031 ns         8031 ns       261096 items_per_second=747.151M/s
ErrorSchemeBoolNoInline            7563 ns         7563 ns       278403 items_per_second=793.376M/s
ErrorSchemeStatusNoInline         16063 ns        16061 ns       130412 items_per_second=373.582M/s
ErrorSchemeResultNoInline         47883 ns        47878 ns        43879 items_per_second=125.319M/s
ErrorSchemeExceptionNoInline       7396 ns         7395 ns       283828 items_per_second=811.316M/s
```
After:
```
ErrorSchemeNoError                 7145 ns         7145 ns       294524 items_per_second=839.806M/s
ErrorSchemeBool                    7565 ns         7564 ns       277756 items_per_second=793.21M/s
ErrorSchemeStatus                  7094 ns         7093 ns       296323 items_per_second=845.898M/s
ErrorSchemeResult                 17428 ns        17424 ns       120180 items_per_second=344.344M/s
ErrorSchemeException               7606 ns         7605 ns       276110 items_per_second=788.972M/s
ErrorSchemeNoErrorNoInline         6631 ns         6630 ns       316267 items_per_second=904.932M/s
ErrorSchemeBoolNoInline            9286 ns         9285 ns       225941 items_per_second=646.179M/s
ErrorSchemeStatusNoInline         15925 ns        15923 ns       131750 items_per_second=376.802M/s
ErrorSchemeResultNoInline         17424 ns        17421 ns       120060 items_per_second=344.403M/s
ErrorSchemeExceptionNoInline       9358 ns         9357 ns       223584 items_per_second=641.213M/s
```

(for some reason, clang refuses to inline the `Result`-returning function...)