@pitrou 2020-02-06
I have tried using `__attribute__((always_inline))` and also passing `-mllvm -inline-threshold=10000`, and clang still doesn't seem to inline that `Result`-returning function. There is definitely something it chokes on.

It may actually be worthwhile trying to make a reproducer and opening an issue on clang?